// packages/quranic-audio-router/src/generated/catalog.data.ts
export const catalog = [
  {
    "reciter": "abdelaziz_sheim",
    "name": {
      "ar": "عبد العزيز سحيم",
      "en": "Abdelaziz Sheim"
    },
// ...
  },
// ...
] as const satisfies readonly GeneratedLeaf[]